Refactor components directory structure
