+++ src/app/components/side-menu/side-menu.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { WelcomeScreenComponent } from '../welcome-screen/welcome-screen.component';
+import { WelcomeScreenComponent } from '../../pages/home-page/welcome-screen/welcome-screen.component';
 import { RouterOutlet } from '@angular/router';
 import { LoadingIndicatorComponent } from '../loading-indicator/loading-indicator.component';
