+++ src/app/pages/home-page/welcome-screen/welcome-screen.component.scss
