+++ src/app/pages/line/direction-selection/direction-selection.component.scss
@@ -1,7 +1,7 @@
+@use '../../../../mixins.scss' as *;
 @use '@angular/material' as mat;
-@use '../../../mixins.scss' as *;
 
-.direction-meanlatency-settings-component {
+.direction-selection-component {
     @include full-height-component;
 
     display: flex;
@@ -22,9 +22,9 @@
             line-height: 28px;
 
             width: fit-content;
-            height: fit-content;
+            height: 36px;
             border-radius: 10px;
-            padding: 4px 10px;
+            padding: 0 10px;
 
             display: flex;
             align-items: center;
@@ -32,10 +32,6 @@
 
             mat-icon {
                 margin-right: 3px;
-            }
-
-            & > span {
-                margin-right: 8px;
             }
         }
 
@@ -46,37 +42,35 @@
     }
 }
 
-.settings-page {
+#direction-selection-buttons {
     display: flex;
     flex-direction: column;
-    padding: 16px;
-}
-
-.setting {
-    padding: 5px;
-    display: flex;
     align-items: center;
-
-    > * {
-        padding: 4px;
-        display: flex;
-        align-items: center;
-    }
+    justify-content: center;
+    padding: 0 16px;
 }
 
-.setting-text {
-    font-weight: 500;
-    display: flex;
-    align-items: center;
-}
-
-.setting-icon {
-    padding-left: 5px;
+.stop-line-button {
+    margin: 7px 5px;
+    padding-left: 15px;
     padding-right: 10px;
-}
+    &.selected {
+        background-color: var(--sys-primary);
+        color: white;
+    }
 
-.button {
-    cursor: pointer;
+    mat-icon {
+        margin-right: 5px;
+    }
+    
+    &.selected {
+        background-color: var(--sys-primary);
+        --mdc-outlined-button-outline-color: var(--sys-primary);
+        color: white;
+    }
+
+    --mdc-outlined-button-label-text-color: #4e544a;
+    --mdc-outlined-button-outline-color: #4e544a;
 }
 
 .navigation-buttons {
