+++ src/app/pages/stop/lines-on-pole-selection/lines-on-pole-selection.component.spec.ts
