+++ src/app/pages/stop/pole-analysis-selection/pole-analysis-selection.component.scss
@@ -1,7 +1,7 @@
-@use '../../../mixins.scss' as *;
+@use '../../../../mixins.scss' as *;
 @use '@angular/material' as mat;
 
-.direction-selection-component {
+.pole-analysis-type-selection-component {
     @include full-height-component;
 
     display: flex;
@@ -22,9 +22,9 @@
             line-height: 28px;
 
             width: fit-content;
-            height: 36px;
+            height: fit-content;
             border-radius: 10px;
-            padding: 0 10px;
+            padding: 4px 10px;
 
             display: flex;
             align-items: center;
@@ -42,35 +42,38 @@
     }
 }
 
-#direction-selection-buttons {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
+#selection-buttons-container {
     padding: 0 16px;
 }
 
-.stop-line-button {
-    margin: 7px 5px;
-    padding-left: 15px;
-    padding-right: 10px;
+.selection-button {
+    border: 1px solid #8D9387;
+    border-radius: 16px;
+    padding: 8px 16px;
+    margin: 10px 0;
+    cursor: pointer;
+
+    .selection-header {
+        width: fit-content;
+        padding: 4px 8px;
+        border-radius: 5px;
+        background-color: var(--sys-primary);
+        color: white;
+        margin-bottom: 7px;
+        font-size: 15px;
+    }
+
     &.selected {
         background-color: var(--sys-primary);
         color: white;
+
+        .selection-header {
+            padding: 4px 0;
+        }
     }
 
-    mat-icon {
-        margin-right: 5px;
-    }
-    
-    &.selected {
-        background-color: var(--sys-primary);
-        --mdc-outlined-button-outline-color: var(--sys-primary);
-        color: white;
-    }
-
-    --mdc-outlined-button-label-text-color: #4e544a;
-    --mdc-outlined-button-outline-color: #4e544a;
+    font-size: 13px;
+    line-height: 15px;
 }
 
 .navigation-buttons {
