+++ src/app/pages/stop/pole-analysis-selection/pole-analysis-selection.component.spec.ts
