+++ src/app/pages/stop/pole-selection/pole-selection.component.html
