+++ src/app/pages/stop/pole-selection/pole-selection.component.scss
@@ -1,13 +1,12 @@
-@use '../../../mixins.scss' as *;
+@use '../../../../mixins.scss' as *;
 @use '@angular/material' as mat;
 
-.pole-analysis-type-selection-component {
+.pole-selection-component {
     @include full-height-component;
 
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-
 
     & > header {
         display: flex;
@@ -42,38 +41,38 @@
     }
 }
 
-#selection-buttons-container {
+#pole-selection-buttons {
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
     padding: 0 16px;
+
+    max-height: 465px;
+    margin-bottom: 7px;
+    overflow-y: auto;
 }
 
-.selection-button {
-    border: 1px solid #8D9387;
-    border-radius: 16px;
-    padding: 8px 16px;
-    margin: 10px 0;
-    cursor: pointer;
+:host ::ng-deep .mat-mdc-button-touch-target {
+    height: unset !important;
+}
 
-    .selection-header {
-        width: fit-content;
-        padding: 4px 8px;
-        border-radius: 5px;
-        background-color: var(--sys-primary);
-        color: white;
-        margin-bottom: 7px;
-        font-size: 15px;
-    }
+.stop-line-button {
+    --mdc-outlined-button-container-height: 35px;
+    margin: 6px 0;
+    margin-right: 10px;
+    padding: 0 12px;
+    width: fit-content;
 
     &.selected {
         background-color: var(--sys-primary);
+        --mdc-outlined-button-outline-color: var(--sys-primary);
         color: white;
-
-        .selection-header {
-            padding: 4px 0;
-        }
     }
 
-    font-size: 13px;
-    line-height: 15px;
+    --mdc-outlined-button-label-text-color: #4e544a;
+    --mdc-outlined-button-outline-color: #4e544a;
 }
 
 .navigation-buttons {
